Connect to SpiceDB in the `hash-graph` binary
