+++ .env
@@ -23,6 +23,7 @@ HASH_KRATOS_PG_DATABASE=kratos
 
 HASH_SPICEDB_VERSION=1.24.0
 HASH_SPICEDB_GRPC_PRESHARED_KEY=secret
+HASH_SPICEDB_HOST=http://127.0.0.1
 HASH_SPICEDB_HTTP_PORT=8443
 HASH_SPICEDB_GRPC_PORT=50051
 HASH_SPICEDB_PG_USER=spicedb
