+++ apps/hash-graph/bin/cli/Cargo.toml
@@ -36,3 +36,4 @@ uuid = "1.4.1"
 
 [features]
 test-server = ["graph/test-server"]
+authorization = []
